+++ satip/constants.py
@@ -61,3 +61,23 @@ SCALER_MAXS = np.array([
 
 HRV_SCALER_MIN = np.array([-1.2278595])
 HRV_SCALER_MAX = np.array([103.90016])
+SAT_VARIABLE_NAMES = (
+    "HRV",
+    "IR_016",
+    "IR_039",
+    "IR_087",
+    "IR_097",
+    "IR_108",
+    "IR_120",
+    "IR_134",
+    "VIS006",
+    "VIS008",
+    "WV_062",
+    "WV_073",
+)
+NATIVE_FILESIZE_MB = 102.210123
+CLOUD_FILESIZE_MB = 3.445185
+RSS_ID = "EO:EUM:DAT:MSG:MSG15-RSS"
+CLOUD_ID = "EO:EUM:DAT:MSG:RSS-CLM"
+SEVIRI_ID = "EO:EUM:DAT:MSG:HRSEVIRI"
+SEVIRI_IODC_ID = "EO:EUM:DAT:MSG:HRSEVIRI-IODC"
