+++ satip/download/goes_download_manager.py
